+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/rule/security/SecurityRulesTest.java
@@ -12,7 +12,7 @@ public class SecurityRulesTest extends SimpleAggregatorTst {
 
     @Override
     public void setUp() {
+        addRule(RULESET, "HardCodedCryptoKey");
         addRule(RULESET, "InsecureCryptoIv");
     }
-
 }