+++ pmd-ui/src/main/java/net/sourceforge/pmd/util/fxdesigner/util/controls/ScopeHierarchyTreeCell.java
@@ -10,11 +10,8 @@ import net.sourceforge.pmd.lang.java.symboltable.MethodNameDeclaration;
 import net.sourceforge.pmd.lang.java.symboltable.VariableNameDeclaration;
 import net.sourceforge.pmd.lang.symboltable.NameDeclaration;
 import net.sourceforge.pmd.lang.symboltable.Scope;
-import net.sourceforge.pmd.util.fxdesigner.MainDesignerController;
 
 import javafx.scene.control.TreeCell;
-import javafx.scene.input.MouseButton;
-import javafx.scene.input.MouseEvent;
 
 
 /**
@@ -25,13 +22,6 @@ import javafx.scene.input.MouseEvent;
  */
 public class ScopeHierarchyTreeCell extends TreeCell<Object> {
 
-    private final MainDesignerController controller;
-
-
-    public ScopeHierarchyTreeCell(MainDesignerController controller) {
-        this.controller = controller;
-    }
-    
     @Override
     protected void updateItem(Object item, boolean empty) {
         super.updateItem(item, empty);
@@ -43,15 +33,6 @@ public class ScopeHierarchyTreeCell extends TreeCell<Object> {
             setText(item instanceof Scope ? getTextForScope((Scope) item)
                                           : getTextForDeclaration((NameDeclaration) item));
         }
-
-        this.addEventHandler(MouseEvent.MOUSE_CLICKED, t -> {
-            if (t.getButton() == MouseButton.PRIMARY 
-                && getTreeView().getSelectionModel().getSelectedItem() == item) {
-                if (item instanceof NameDeclaration) {
-                    controller.onNodeItemSelected(((NameDeclaration) item).getNode());
-                }
-            }
-        });
     }